worked on component2 and created remove_duplicates_logic.py , added duplicate characters functionality
